fix(openai_dart): Consolidate duplicate copyWith sentinels and fix doc comments
Delete the responses-specific copy_with_sentinel.dart (which used a different
private class) and redirect all importers to the canonical one under
models/common/. Also remove doc comments referencing the non-existent
`conversation` parameter and `ResponseConversation` class.
