+++ packages/openai_dart/lib/src/client/openai_client.dart
@@ -945,7 +945,6 @@ class OpenAIClient {
   ///   CreateResponseRequest(
   ///     model: 'gpt-4o',
   ///     input: ResponseInput.text('Continue our conversation'),
-  ///     conversation: ResponseConversation.id(conversation.id),
   ///   ),
   /// );
   ///
