+++ packages/openai_dart/lib/src/models/conversations/conversations.dart
@@ -29,7 +29,6 @@
 ///   CreateResponseRequest(
 ///     model: 'gpt-4o',
 ///     input: ResponseInput.text('Continue our conversation'),
-///     conversation: ResponseConversation.id(conversation.id),
 ///   ),
 /// );
 ///
