+++ packages/openai_dart/lib/src/models/responses/common/common.dart
@@ -1,5 +1,4 @@
 /// Common utilities for the Responses API models.
 library;
 
-export 'copy_with_sentinel.dart';
 export 'equality_helpers.dart';
