--- packages/openai_dart/lib/src/models/responses/common/copy_with_sentinel.dart
@@ -1,7 +1,0 @@
-/// Sentinel object used in copyWith methods to distinguish between
-/// explicit null values and unset values.
-const Object unsetCopyWithValue = _CopyWithSentinel();
-
-class _CopyWithSentinel {
-  const _CopyWithSentinel();
-}
