+++ packages/openai_dart/lib/src/models/responses/items/item.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 import 'package:meta/meta.dart';
 
-import '../common/copy_with_sentinel.dart';
+import '../../common/copy_with_sentinel.dart';
 import '../common/equality_helpers.dart';
 import '../config/function_call_status.dart';
 import '../config/item_status.dart';
