+++ packages/openai_dart/lib/src/resources/conversations_resource.dart
@@ -23,7 +23,6 @@ import 'base_resource.dart';
 ///   CreateResponseRequest(
 ///     model: 'gpt-4o',
 ///     input: ResponseInput.text('Continue our conversation'),
-///     conversation: ResponseConversation.id(conversation.id),
 ///   ),
 /// );
 ///
